+++ lib/crawler/worker.ex
@@ -3,11 +3,15 @@ defmodule Crawler.Worker do
   Handles the crawl tasks.
   """
 
+  require Logger
+
   alias Crawler.{Fetcher, Store, Store.Page}
 
   use GenServer
 
   def init(args) do
+    :timer.send_after(args[:timeout], :stop)
+
     {:ok, args}
   end
 
@@ -15,7 +19,9 @@ defmodule Crawler.Worker do
   Runs the worker that casts data to itself to kick off the crawl workflow.
   """
   def run(opts) do
-    {:ok, pid} = GenServer.start_link(__MODULE__, opts)
+    Logger.info("Running worker with opts: #{inspect(opts)}")
+
+    {:ok, pid} = GenServer.start_link(__MODULE__, opts, hibernate_after: 0)
 
     GenServer.cast(pid, opts)
   end
@@ -27,15 +33,20 @@ defmodule Crawler.Worker do
   - `Crawler.Parser.parse/1` (or a custom parser)
   """
   def handle_cast(_req, state) do
+    Logger.info("Running worker with opts: #{inspect(state)}")
+
     state
     |> Fetcher.fetch()
     |> state[:parser].parse()
     |> mark_processed()
 
-    {:noreply, state}
+    {:noreply, state, :hibernate}
   end
 
-  @doc false
+  def handle_info(:stop, state) do
+    {:stop, :normal, state}
+  end
+
   def handle_info(_msg, state) do
     {:noreply, state}
   end